requirements.txt
streamlit==1.38.0
anthropic==0.50.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.4.0
pandas==2.2.3
langchain==0.3.24
langchain-core==0.3.56
tiktoken==0.9.0
chromadb==0.5.5
watchdog==4.0.2
reportlab==4.4.0
